requirements.txt
opencv-python
face-recognition
numpy
pandas
Pillow